backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Gutenberg Analyzer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=3.0.1",
    "flask>=2.2.5",
    "flask-cors>=5.0.0",
    "groq>=0.31.0",
    "requests>=2.31.0",
    "openai<=1.99.9", # Pin due to litellm incompatibility with >=1.100.0
    "fastapi>=0.116.1",
    "orjson>=3.11.2",
    "litellm>=1.75.3",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (slow, requires network)",
]
